pyproject.toml
[project]
name = "ai-agent-demo"
version = "0.1.0"
description = "AI Agent Demo with AgentScope"
requires-python = ">=3.10"
dependencies = [
    "agentscope>=0.1.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
]
